test && fix errors

+++ www/api/tweets.py
@@ -13,7 +13,7 @@ def write_tweet(content, photos, visibility=Visibility.All):
         raise APIError('内容不能为空')
     if not isinstance(photos, list):
         photos = []
-    if photos.__len__ > 3:
+    if photos.__len__() > 3:
         raise APIError('图片不能超过三张')
 
     uid = session['uid']
@@ -49,36 +49,41 @@ def reply(target_tweet, content, visibility=Visibility.All):
 
     db.session.add(reply)
     db.session.commit()
+    return reply
 
 
 def get_friends_tweets(offset=0, limit=10, later_than=0):
     current_uid = session['uid']
     
     friends = api.friends.get_friends()
-    tweets = [ Tweet.query.filter(user==item.to, created_at>=later_than).all() for item in friends ]
+    tweets = [ Tweet.query.filter(Tweet.user==item.to, Tweet.created_at>=later_than).all() for item in friends ]
     res = []
     for i in tweets:
         res.extend(i)
     res.extend(Tweet.query.filter_by(user=current_uid).all())
-    return res[offset: offset+limit].sort(key=lambda tweet: tweet.created_at, reverse=True)
+    res.sort(key=lambda tweet: tweet.created_at, reverse=True)
+
+    return res[offset: offset+limit]
 
 
 def get_users_tweets(uid, offset=0, limit=10, later_than=0):
     current_uid = session['uid']
 
-    friends = [ friend.to for friend in Friend.query.filter(user==current_uid, agree=True).all() ]
-    tweets = Tweet.query.filter(user==uid, created_at>=later_than).order_by(Tweet.created_at.desc()).all()
+    friends = [ friend.to for friend in Friend.query.filter(Friend.user==current_uid, Friend.agree==True).all() ]
+    tweets = Tweet.query.filter(Tweet.user==uid, Tweet.created_at>=later_than).order_by(Tweet.created_at.desc()).all()
     for item in tweets:
         if item.visibility == Visibility.FriendsOnly and not uid in friends:
             tweets.remove(item)
-    return tweets[offset : offset+limit].sort(key=lambda tweet: tweet.created_at, reverse=True)
+    tweets.sort(key=lambda tweet: tweet.created_at, reverse=True)
+
+    return tweets[offset : offset+limit]
 
 
 def get_replies(tweet_id, offset=0, limit=10, later_than=0):
     current_uid = session['uid']
 
     this_tweet = Tweet.query.filter_by(id=tweet_id).first()
-    replies = Reply.query.filter(target==tweet_id, created_at>=later_than).all()
+    replies = Reply.query.filter(Reply.target==tweet_id, Reply.created_at>=later_than).all()
     for item in replies:
         if item.visibility == Visibility.Mutual and ( current_uid != item.user or current_uid != this_tweet.user):
             replies.remove(item)
@@ -92,6 +97,8 @@ def remove_tweet(id):
     current_uid = session['uid']
 
     tweet = Tweet.query.filter_by(id=id).first()
+    if not tweet:
+        raise APIError('此说说不存在')
     if not tweet.user == current_uid:
         raise APIError('您没有权限删除此说说。必须是自己发的才行。')
     replies = Reply.query.filter_by(target=tweet.id).all()
@@ -107,7 +114,12 @@ def remove_reply(id):
     current_uid = session['uid']
 
     reply = Reply.query.filter_by(id=id).first()
+    if not reply:
+        raise APIError('此回复不存在')
     tweet = Tweet.query.filter_by(id=reply.target).first()
+    if not tweet:
+        raise APIError('回复对应的说说不存在。')
+
     if reply.user == current_uid or tweet.user == current_uid:
         db.session.delete(reply)
         db.session.commit()
